moves the new DimcatSchema and DimcatConfig to base.py and adapts DimcatObject accordingly
The playground file test_config.py has been renamed to test_base.py and a first proper unittest, TestDimcatObject, has been introduced.
